% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binspp-package.R
\docType{package}
\name{binspp-package}
\alias{binspp-package}
\title{binspp: Bayesian Inference for Neyman-Scott Point Processes}
\description{
The Bayesian MCMC estimation of parameters for Thomas-type cluster point process with various inhomogeneities. It allows for inhomogeneity in (i) distribution of parent points, (ii) mean number of points in a cluster, (iii) cluster spread. The package also allows for the Bayesian MCMC algorithm for the homogeneous generalized Thomas process. The cluster size is allowed to have a variance that is greater or less than the expected value (cluster sizes are over or under dispersed). Details are described in Dvořák, Remeš, Beránek & Mrkvička (2022) <arXiv: 10.48550/arXiv.2205.07946>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tomasmrkvicka/binspp}
}

}
\author{
\strong{Maintainer}: Remes Radim \email{inrem@jcu.cz}

Authors:
\itemize{
  \item Mrkvicka Tomas \email{mrkvicka.toma@gmail.com}
  \item Dvorak Jiri \email{dvorak@karlin.mff.cuni.cz}
  \item Beranek Ladislav \email{beranek@jcu.cz}
}

Other contributors:
\itemize{
  \item Park Jaewoo \email{jwpark88@yonsei.ac.kr} [contributor]
  \item Lee Sujeong \email{dltnwjd2304@gmail.com} [contributor]
}

}
\keyword{internal}
