% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binspp-package.R
\docType{package}
\name{binspp}
\alias{binspp}
\title{binspp: Bayesian inference for Neyman-Scott point processes}
\description{
This package introduces the Bayesian MCMC algorithm for generalized Thomas process, 
   allowing cluster size to have a variance that is greater or less than the expected value 
   (cluster sizes are over- or under-dispersed).
}
\details{
Simulation of generalized Thomas process, Bayesian MCMC estimation of parameters for generalized Thomas process.
}
\note{
LinkingTo: Rcpp,RcppArmadillo, RcppEigen
    License: GPL-3
    Encoding: UTF-8
}
\author{
Tomas Mrkvicka <mrkvicka.toma@gmail.com> (author), Ladislav Beranek <beranek@jcu.cz> (maintainer), Radim Remes <inrem@jcu.cz> (maintainer)
}
