% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coeff}
\alias{coeff}
\title{Calculate parameters for Birth and Death Interaction likelihood functions.}
\usage{
coeff(theta)
}
\arguments{
\item{theta}{A vector of theta1 and theta2.}
}
\value{
A vector of r1 and r2.
}
\description{
Calculates r1 and r2, which serve as distance thresholds for Birth and Death Interaction likelihood functions.
}
