% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_refor.R
\docType{data}
\name{cov_refor}
\alias{cov_refor}
\title{Distance to the reforestration polygon}
\format{
An object of class \code{im} with 208 rows and 374 columns.
}
\usage{
cov_refor
}
\description{
Covariate for data trees_N4.
}
\keyword{datasets}
