% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_reserv.R
\docType{data}
\name{cov_reserv}
\alias{cov_reserv}
\title{Distance to the reservoir}
\format{
An object of class \code{im} with 208 rows and 374 columns.
}
\usage{
cov_reserv
}
\description{
Covariate for data trees_N4.
}
\keyword{datasets}
