% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_slope.R
\docType{data}
\name{cov_slope}
\alias{cov_slope}
\title{Slope of the area}
\format{
An object of class \code{im} with 208 rows and 374 columns.
}
\usage{
cov_slope
}
\description{
Covariate for data trees_N4.
}
\keyword{datasets}
