% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_tdensity.R
\docType{data}
\name{cov_tdensity}
\alias{cov_tdensity}
\title{Trees density}
\format{
An object of class \code{im} with 208 rows and 374 columns.
}
\usage{
cov_tdensity
}
\description{
Covariate for data trees_N4.
}
\keyword{datasets}
