% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_tmi.R
\docType{data}
\name{cov_tmi}
\alias{cov_tmi}
\title{Topographic moisture index}
\format{
An object of class \code{im} with 208 rows and 374 columns.
}
\usage{
cov_tmi
}
\description{
Covariate for data trees_N4.
}
\keyword{datasets}
