% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estgtp.R
\name{estgtp}
\alias{estgtp}
\title{estgtp}
\usage{
estgtp(
  X,
  kappa0 = exp(a_kappa + ((b_kappa^2)/2)),
  omega0 = exp(a_omega + ((b_omega^2)/2)),
  lambda0 = (l_lambda + u_lambda)/2,
  theta0 = exp(a_theta + ((b_theta^2)/2)),
  skappa,
  somega,
  dlambda,
  stheta,
  smove,
  a_kappa,
  b_kappa,
  a_omega,
  b_omega,
  l_lambda,
  u_lambda,
  a_theta,
  b_theta,
  iter = 5e+05,
  plot.step = 1000,
  save.step = 1000,
  filename
)
}
\arguments{
\item{omega0}{Initial value for omega, by default it will be set as expectation of prior for omega.}

\item{lambda0}{Initial value for lambda, by default it will be set as expectation of prior for lambda.}

\item{theta0}{Initial value for theta, by default it will be set as expectation of prior for theta.}

\item{skappa}{variability of proposal for kappa: second parameter of log-normal distribution}

\item{somega}{variability of proposal for omega: second parameter of log-normal distribution}

\item{dlambda}{variability of proposal for lambda: half of range of uniform distribution}

\item{stheta}{variability of proposal for theta: second parameter of log-normal distribution}

\item{smove}{variability of proposal for moving center point: SD of normal distribution}

\item{X:}{A point pattern dataset (object of class "ppp") to which the model should be fitted.}

\item{kappa0:}{Initial value for kappa, by default it will be set as expectation of prior for kappa.}

\item{a_kappa:}{First parameter of prior distribution for kappa, which is log-normal distribution.}

\item{b_kappa:}{Second parameter of prior distribution for kappa, which is log-normal distribution.}

\item{a_omega:}{First parameter of prior distribution for omega, which is log-normal distribution.}

\item{b_omega:}{Second parameter of prior distribution for omega, which is log-normal distribution.}

\item{l_lambda:}{First parameter of prior distribution for lambda, which is uniform distribution.}

\item{u_lambda:}{Second parameter of prior distribution for lambda, which is uniform distribution.}

\item{a_theta:}{First parameter of prior distribution for theta, which is log-normal distribution.}

\item{b_theta:}{Second parameter of prior distribution for theta, which is log-normal distribution.}

\item{iter:}{Number of iterations of MCMC.}

\item{plot.step:}{Step for the graph plotting}

\item{save.step:}{Step for the parameters saving. The file must be specified or has to be set to larger than iter.}

\item{filename:}{The name of the output RDS file}
}
\value{
The output is an estimated MCMC chain of parameters, centers and connections.
}
\description{
Bayesian MCMC estimation of parameters of generalized Thomas process.
}
