% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estgtpr.R
\name{estgtpr}
\alias{estgtpr}
\title{estgtpr}
\usage{
estgtpr(est, discard = 100, step = 10)
}
\arguments{
\item{est:}{Output from estgtp function.}

\item{discard:}{Number of iterations to be discarded as burn in for the estimation.}

\item{step:}{Every step iteration is taken in the parameter estimation.}
}
\value{
median and quantile values and plots (kappa, omega, lambda, theta)
}
\description{
Calculates median values for kappa, omega, lambda, theta; calculates 2.5 and 97.5 quantile and draws trace plots
}
