% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgtp.R
\name{rgtp}
\alias{rgtp}
\title{RGTP}
\usage{
rgtp(
  kappa,
  omega,
  lambda,
  theta,
  win = owin(c(0, 1), c(0, 1)),
  nsim = 1,
  expand = 4 * omega,
  C = NULL
)
}
\arguments{
\item{kappa:}{Intensity of the Poisson process of cluster centres. A single positive number.}

\item{omega:}{Standard deviation of random displacement (along each coordinate axis) of a point from its cluster centre.}

\item{lambda:}{Shape parameter of Generalized Poisson distribution for number of points in a cluster.}

\item{theta:}{Location parameter of Generalized Poisson distribution for number of points in a cluster.}

\item{win:}{Window in which to simulate the pattern. An object of class "owin" or something acceptable to as.owin.}

\item{nsim:}{Number of simulations.}

\item{expand:}{The size of expansion of window to simulate the centers of clusters.}

\item{C:}{Process of center points.}
}
\value{
A list(X, C), where X is Generalized Poisson process, and C is Process of cluster centers for Generalized Poisson process.
}
\description{
Simulation of generalized Thomas process.
}
