% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_top_N4.R
\docType{data}
\name{y_top_N4}
\alias{y_top_N4}
\title{Vertical corners for trees_N4 dataset}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
y_top_N4
}
\description{
The vector of top vertical corners of rectangles forming observation
  window for trees_N4 dataset.
}
\keyword{datasets}
